+++ system/platform-core/src/test/java/org/platformlambda/core/models/SimplePoJo.java
@@ -16,16 +16,11 @@
 
  */
 
-package org.platformlambda.core.mock
+package org.platformlambda.core.models;
 
-import org.platformlambda.common.JacksonSerializer
-import org.platformlambda.core.models.SimplePoJo
-import org.platformlambda.core.annotations.PreLoad
-import org.platformlambda.core.models.KotlinLambdaFunction
+public class SimplePoJo {
 
-@PreLoad(route = "custom.serializer.service.kotlin", customSerializer = JacksonSerializer::class)
-class CustomSerializerKotlinService : KotlinLambdaFunction<SimplePoJo, SimplePoJo> {
-    override suspend fun handleEvent(headers: Map<String, String>, input: SimplePoJo, instance: Int): SimplePoJo {
-        return input
-    }
+    public String name;
+    public String address;
+    public int telephone;
 }