+++ CHANGELOG.md
@@ -8,6 +8,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.1.5, 1/14/2024
+
+### Added
+
+N/A
+
+### Removed
+
+Custom Map/List deserializers
+
+### Changed
+
+Apply ToNumberPolicy.LONG_OR_DOUBLE to GSON serialization of untyped numbers
+in hash map.
+
+---
 ## Version 4.1.4, 1/13/2024
 
 ### Added